Data.X509

Description

Read/Write X509 Certificate, CRL and their signed equivalents.

Follows RFC5280 / RFC6818

Types

type SignedCertificate = SignedExact Certificate

A Signed Certificate

type SignedCRL = SignedExact CRL

A Signed CRL

data Certificate

X.509 Certificate type.

This type doesn't include the signature, it's describe in the RFC as tbsCertificate.

Constructors

Certificate

Fields

certVersion :: Int

Version

certSerial :: Integer

Serial number

certSignatureAlg :: SignatureALG

Signature algorithm

certIssuerDN :: DistinguishedName

Issuer DN

certValidity :: (DateTime, DateTime)

Validity period (UTC)

certSubjectDN :: DistinguishedName

Subject DN

certPubKey :: PubKey

Public key

certExtensions :: Extensions

Extensions

data PubKey

Public key types known and used in X.509

Constructors

PubKeyRSA PublicKey
RSA public key

PubKeyDSA PublicKey
DSA public key

PubKeyDH (Integer, Integer, Integer, Maybe Integer, ([Word8], Integer))
DH format with (p,g,q,j,(seed,pgenCounter))

PubKeyEC PubKeyEC
EC public key

PubKeyX25519 PublicKey
X25519 public key

PubKeyX448 PublicKey
X448 public key

PubKeyEd25519 PublicKey
Ed25519 public key

PubKeyEd448 PublicKey
Ed448 public key

PubKeyUnknown OID ByteString
unrecognized format

data PubKeyEC

Elliptic Curve Public Key

TODO: missing support for binary curve.

Constructors

PubKeyEC_Prime

Fields

pubkeyEC_pub :: SerializedPoint

pubkeyEC_a :: Integer

pubkeyEC_b :: Integer

pubkeyEC_prime :: Integer

pubkeyEC_generator :: SerializedPoint

pubkeyEC_order :: Integer

pubkeyEC_cofactor :: Integer

pubkeyEC_seed :: Integer

PubKeyEC_Named

Fields

pubkeyEC_name :: CurveName

pubkeyEC_pub :: SerializedPoint

newtype SerializedPoint

Serialized Elliptic Curve Point

Constructors

SerializedPoint ByteString

data PrivKey

Private key types known and used in X.509

Constructors

PrivKeyRSA PrivateKey
RSA private key

PrivKeyDSA PrivateKey
DSA private key

PrivKeyEC PrivKeyEC
EC private key

PrivKeyX25519 SecretKey
X25519 private key

PrivKeyX448 SecretKey
X448 private key

PrivKeyEd25519 SecretKey
Ed25519 private key

PrivKeyEd448 SecretKey
Ed448 private key

data PrivKeyEC

Elliptic Curve Private Key

TODO: missing support for binary curve.

Constructors

PrivKeyEC_Prime

Fields

privkeyEC_priv :: Integer

privkeyEC_a :: Integer

privkeyEC_b :: Integer

privkeyEC_prime :: Integer

privkeyEC_generator :: SerializedPoint

privkeyEC_order :: Integer

privkeyEC_cofactor :: Integer

privkeyEC_seed :: Integer

PrivKeyEC_Named

Fields

privkeyEC_name :: CurveName

privkeyEC_priv :: Integer

pubkeyToAlg :: PubKey -> PubKeyALG

Convert a Public key to the Public Key Algorithm type

privkeyToAlg :: PrivKey -> PubKeyALG

Convert a Private key to the Public Key Algorithm type

data HashALG

Hash Algorithm

Constructors

HashMD2

HashMD5

HashSHA1

HashSHA224

HashSHA256

HashSHA384

HashSHA512

data PubKeyALG

Public Key Algorithm

Constructors

PubKeyALG_RSA
RSA Public Key algorithm

PubKeyALG_RSAPSS
RSA PSS Key algorithm (RFC 3447)

PubKeyALG_DSA
DSA Public Key algorithm

PubKeyALG_EC
ECDSA & ECDH Public Key algorithm

PubKeyALG_X25519
ECDH 25519 key agreement

PubKeyALG_X448
ECDH 448 key agreement

PubKeyALG_Ed25519
EdDSA 25519 signature algorithm

PubKeyALG_Ed448
EdDSA 448 signature algorithm

PubKeyALG_DH
Diffie Hellman Public Key algorithm

PubKeyALG_Unknown OID
Unknown Public Key algorithm

data SignatureALG

Signature Algorithm, often composed of a public key algorithm and a hash algorithm. For some signature algorithms the hash algorithm is intrinsic to the public key algorithm and is not needed in the data type.

Constructors

SignatureALG HashALG PubKeyALG

SignatureALG_IntrinsicHash PubKeyALG

SignatureALG_Unknown OID

class Extension a where

Extension class.

each extension have a unique OID associated, and a way to encode and decode an ASN1 stream.

Errata: turns out, the content is not necessarily ASN1, it could be data that is only parsable by the extension e.g. raw ascii string. Add method to parse and encode with ByteString

Minimal complete definition

extOID, extHasNestedASN1, extEncode, extDecode

Methods

extOID :: a -> OID

extHasNestedASN1 :: Proxy a -> Bool

extEncode :: a -> [ASN1]

extDecode :: [ASN1] -> Either String a

extDecodeBs :: ByteString -> Either String a

extEncodeBs :: a -> ByteString

Common extension usually found in x509v3

data ExtBasicConstraints

Basic Constraints

Constructors

ExtBasicConstraints Bool (Maybe Integer)

data ExtKeyUsage

Describe key usage

Constructors

ExtKeyUsage [ExtKeyUsageFlag]

data ExtKeyUsageFlag

key usage flag that is found in the key usage extension field.

Constructors

KeyUsage_digitalSignature

KeyUsage_nonRepudiation

KeyUsage_keyEncipherment

KeyUsage_dataEncipherment

KeyUsage_keyAgreement

KeyUsage_keyCertSign

KeyUsage_cRLSign

KeyUsage_encipherOnly

KeyUsage_decipherOnly

data ExtExtendedKeyUsage

Extended key usage extension

Constructors

ExtExtendedKeyUsage [ExtKeyUsagePurpose]

data ExtKeyUsagePurpose

Key usage purposes for the ExtendedKeyUsage extension

Constructors

KeyUsagePurpose_ServerAuth

KeyUsagePurpose_ClientAuth

KeyUsagePurpose_CodeSigning

KeyUsagePurpose_EmailProtection

KeyUsagePurpose_TimeStamping

KeyUsagePurpose_OCSPSigning

KeyUsagePurpose_Unknown OID

data ExtSubjectKeyId

Provide a way to identify a public key by a short hash.

Constructors

ExtSubjectKeyId ByteString

data ExtSubjectAltName

Provide a way to supply alternate name that can be used for matching host name.

Constructors

ExtSubjectAltName [AltName]

data ExtAuthorityKeyId

Provide a mean to identify the public key corresponding to the private key used to signed a certificate.

Constructors

ExtAuthorityKeyId ByteString

data ExtCrlDistributionPoints

Identify how CRL information is obtained

Constructors

ExtCrlDistributionPoints [DistributionPoint]

data ExtNetscapeComment

Constructors

ExtNetscapeComment ByteString

data AltName

Different naming scheme use by the extension.

Not all name types are available, missing: otherName x400Address directoryName ediPartyName registeredID

Constructors

AltNameRFC822 String

AltNameDNS String

AltNameURI String

AltNameIP ByteString

AltNameXMPP String

AltNameDNSSRV String

data DistributionPoint

Distribution point as either some GeneralNames or a DN

Constructors

DistributionPointFullName [AltName]

DistributionNameRelative DistinguishedName

data ReasonFlag

Reason flag for the CRL

Constructors

Reason_Unused

Reason_KeyCompromise

Reason_CACompromise

Reason_AffiliationChanged

Reason_Superseded

Reason_CessationOfOperation

Reason_CertificateHold

Reason_PrivilegeWithdrawn

Reason_AACompromise

Accessor turning extension into a specific one

extensionGet :: Extension a => Extensions -> Maybe a

Get a specific extension from a lists of raw extensions

extensionGetE :: Extension a => Extensions -> Maybe (Either String a)

Get a specific extension from a lists of raw extensions

extensionDecode :: forall a. Extension a => ExtensionRaw -> Maybe (Either String a)

Try to decode an ExtensionRaw.

If this function return: * Nothing, the OID doesn't match * Just Left, the OID matched, but the extension couldn't be decoded * Just Right, the OID matched, and the extension has been succesfully decoded

extensionEncode :: forall a. Extension a => Bool -> a -> ExtensionRaw

Encode an Extension to extensionRaw

data ExtensionRaw

An undecoded extension

Constructors

ExtensionRaw

Fields

extRawOID :: OID

OID of this extension

extRawCritical :: Bool

if this extension is critical

extRawContent :: ByteString

undecoded content

tryExtRawASN1 :: ExtensionRaw -> Either String [ASN1]

extRawASN1 :: ExtensionRaw -> [ASN1]

Deprecated: use tryExtRawASN1 instead

newtype Extensions

a Set of ExtensionRaw

Constructors

Extensions (Maybe [ExtensionRaw])

Certificate Revocation List (CRL)

data CRL

Describe a Certificate revocation list

Constructors

CRL

Fields

crlVersion :: Integer

crlSignatureAlg :: SignatureALG

crlIssuer :: DistinguishedName

crlThisUpdate :: DateTime

crlNextUpdate :: Maybe DateTime

crlRevokedCertificates :: [RevokedCertificate]

crlExtensions :: Extensions

data RevokedCertificate

Describe a revoked certificate identifiable by serial number.

Constructors

RevokedCertificate

Fields

revokedSerialNumber :: Integer

revokedDate :: DateTime

revokedExtensions :: Extensions

Naming

newtype DistinguishedName

A list of OID and strings.

Constructors

DistinguishedName

Fields

getDistinguishedElements :: [(OID, ASN1CharacterString)]

data DnElement

Elements commonly available in a DistinguishedName structure

Constructors

DnCommonName
CN

DnCountry
Country

DnOrganization
O

DnOrganizationUnit
OU

DnEmailAddress
Email Address (legacy)

data ASN1CharacterString

ASN1 Character String with encoding

Constructors

ASN1CharacterString

Fields

characterEncoding :: ASN1StringEncoding

getCharacterStringRawData :: ByteString

getDnElement :: DnElement -> DistinguishedName -> Maybe ASN1CharacterString

Try to get a specific element in a DistinguishedName structure

Certificate Chain

newtype CertificateChain

A chain of X.509 certificates in exact form.

Constructors

CertificateChain [SignedExact Certificate]

newtype CertificateChainRaw

Represent a chain of X.509 certificates in bytestring form.

Constructors

CertificateChainRaw [ByteString]

marshall between CertificateChain and CertificateChainRaw

decodeCertificateChain :: CertificateChainRaw -> Either (Int, String) CertificateChain

Decode a CertificateChainRaw into a CertificateChain if every raw certificate are decoded correctly, otherwise return the index of the failed certificate and the error associated.

encodeCertificateChain :: CertificateChain -> CertificateChainRaw

Convert a CertificateChain into a CertificateChainRaw

Signed types and marshalling

data (Show a, Eq a, ASN1Object a) => Signed a

Represent a signed object using a traditional X509 structure.

When dealing with external certificate, use the SignedExact structure not this one.

Constructors

Signed

Fields

signedObject :: a

Object to sign

signedAlg :: SignatureALG

Signature Algorithm used

signedSignature :: ByteString

Signature as bytes

data (Show a, Eq a, ASN1Object a) => SignedExact a

Represent the signed object plus the raw data that we need to keep around for non compliant case to be able to verify signature.

getSigned :: SignedExact a -> Signed a

get the decoded Signed data

getSignedData :: (Show a, Eq a, ASN1Object a) => SignedExact a -> ByteString

Get the signed data for the signature

objectToSignedExact

Arguments

:: (Show a, Eq a, ASN1Object a)

=> (ByteString -> (ByteString, SignatureALG, r))
signature function

-> a
object to sign

-> (SignedExact a, r)

Transform an object into a SignedExact object

objectToSignedExactF

Arguments

:: (Functor f, Show a, Eq a, ASN1Object a)

=> (ByteString -> f (ByteString, SignatureALG))
signature function

-> a
object to sign

-> f (SignedExact a)

A generalization of objectToSignedExact where the signature function runs in an arbitrary functor. This allows for example to sign using an algorithm needing random values.

encodeSignedObject :: SignedExact a -> ByteString

The raw representation of the whole signed structure

decodeSignedObject :: (Show a, Eq a, ASN1Object a) => ByteString -> Either String (SignedExact a)

Try to parse a bytestring that use the typical X509 signed structure format

Parametrized Signed accessor

getCertificate :: SignedCertificate -> Certificate

Get the Certificate associated to a SignedCertificate

getCRL :: SignedCRL -> CRL

Get the CRL associated to a SignedCRL

decodeSignedCertificate :: ByteString -> Either String SignedCertificate

Try to decode a bytestring to a SignedCertificate

decodeSignedCRL :: ByteString -> Either String SignedCRL

Try to decode a bytestring to a SignedCRL

Hash distinguished names related function

hashDN :: DistinguishedName -> ByteString

Make an OpenSSL style hash of distinguished name

OpenSSL algorithm is odd, and has been replicated here somewhat. only lower the case of ascii character.

hashDN_old :: DistinguishedName -> ByteString

Create an openssl style old hash of distinguished name